Reactor telemetry appenders to reduce duplicated code
